Use table definitions from chimedb.data_index
