--- alpenhorn_chime/info/__init__.py
@@ -1,7 +1,0 @@
-"""CHIME info classes."""
-
-from .cal import cal_info_class
-from .corr import CorrAcqInfo, CorrFileInfo
-from .hfb import HFBAcqInfo, HFBFileInfo
-from .rawadc import RawadcAcqInfo, RawadcFileInfo
-from .weather import WeatherFileInfo
